requirements.txt
selenium>=4.15.2
webdriver-manager>=4.0.1 
streamlit>=1.45.1